Tyler: Code refresh to keep up with main repo
