+++ MetaScanner/Properties/AssemblyInfo.cs
@@ -32,5 +32,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("2.0.5.1018")]
-[assembly: AssemblyFileVersion("2.0.5.1018")]
+[assembly: AssemblyVersion("2.0.6.1019")]
+[assembly: AssemblyFileVersion("2.0.6.1019")]
