+++ MetaScanner/UI/Mini/ScannerViewMini.cs
@@ -29,7 +29,7 @@ namespace inSSIDer.UI.Mini
 {
     public partial class ScannerViewMini : UserControl
     {
-        private ScannerN _sc;
+        private ScanController _sc;
         //selection ignore flag - set to makes it ignore the first the auto-select when a new rows is added
         private bool _ignoreSelection = true;
 
@@ -45,7 +45,7 @@ namespace inSSIDer.UI.Mini
             InitializeComponent();
         }
 
-        public void SetScanner(ref ScannerN scanner)
+        public void SetScanner(ref ScanController scanner)
         {
             _sc = scanner;
             _sc.Cache.DataReset += Cache_DataReset;
